extracted listview to widgets. Added docs
